Configure shutdown watchdog to 20min
There were reports of the shutdown watchdog firing in OpenBMC #522 which caused
the flashing process to fail. Andrew Geissler suggested he reached ~77%
before the watchdog bit, which by default is configured at 10 minutes.
Back-of-the-envelope calculations suggest it would've taken
approximately 13 minutes to complete the flash, therefore bump the
number to 20min to give ourselves a bit of head-room.

https://github.com/openbmc/openbmc/issues/522

Change-Id: Ie404682dbccc0225d939c382c107610daf9a41ac
Signed-off-by: Peter Lin <peter.cc.lin@mail.foxconn.com>